ai-services-testing/requirements.txt
openai-whisper==20230314
pytest>=7.0.0
pydantic>=2.0.0
numpy>=1.24.0
scipy>=1.10.0
requests>=2.25.0
pytest-cov>=4.0.0